% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessing.R
\name{checkPreloadKinaseTable}
\alias{checkPreloadKinaseTable}
\title{Print a summary kinase/phosphatase table}
\usage{
checkPreloadKinaseTable(Abbrev = NULL)
}
\arguments{
\item{Abbrev}{(Optional) The function returns the available IDs of the listed abbreviated species. If left as NULL, the function will output the abbreviation and full species name of the 25 species for which inases and/or phosphatases are available.}
}
\value{
data.frame
}
\description{
Check the preloaded kinase/phosphatase table.
}
\examples{
\dontrun{
## Print the 25 Abbreviations and Species.
checkPreloadKinaseTable()
## Print the IDs of the provided species
checkPreloadKinaseTable(Abbrev = "Ath")
}
}
