% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessing.R
\name{confirmColumnNames}
\alias{confirmColumnNames}
\title{Identify the Columns needed for Data Processing}
\usage{
confirmColumnNames(
  rawMaxQuant = rawMaxQuant,
  positionCol = positionCol,
  reverseCol = reverseCol,
  localizationProbCol = localizationProbCol,
  potentialContaminationCol = potentialContaminationCol,
  aminoAcidCol = aminoAcidCol,
  uniqueIDCol = uniqueID,
  seqWindowIDCol = seqWindowID,
  fastaIDCol = NULL
)
}
\arguments{
\item{rawMaxQuant}{(Required). Loaded raw MaxQuant data}

\item{positionCol}{(Required). List the column name with the phosphorylation position.}

\item{reverseCol}{(Required). List the column with the reverse information.}

\item{localizationProbCol}{(Required). List the column name with the localization probability.}

\item{potentialContaminationCol}{(Required). List the column name with the contamination information.}

\item{aminoAcidCol}{(Required). List the column name with the amino acid.}

\item{uniqueIDCol}{(Required). List the column name with the protein  for the peptides.}

\item{seqWindowIDCol}{(Required). List the column name with the sequence window.}

\item{fastaIDCol}{(Optional). List the column name with the FASTA header.}
}
\value{
A list of data.frames
}
\description{
The function selects the correct columns. Please note that the column name is case-sensitive.
}
\examples{
\dontrun{
## Loading One Condition Data
data("oneConditionExample")
## Identify the Columns needed for Data Processing
identifiedCols <- confirmColumnNames(rawMaxQuant = oneConditionExample,
                                	    positionCol = "Position",
                                	    reverseCol = "Reverse",
                                     localizationProbCol = "Localization prob",
                                     potentialContaminationCol = "Potential contaminant",
                                     aminoAcidCol = "Amino acid",
                                	    uniqueIDCol = "Protein",
                                     seqWindowIDCol = "Sequence window",
                    	                fastaIDCol = "Fasta headers")
}
}
