% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessing.R
\name{confirmIntensityColumns}
\alias{confirmIntensityColumns}
\title{Identify the Intensity Columns}
\usage{
confirmIntensityColumns(
  rawMaxQuant = rawMaxQuant,
  intensityPattern = intensityPattern,
  filterCon = NULL,
  filterTime = NULL,
  filterRep = NULL,
  verbose = TRUE
)
}
\arguments{
\item{rawMaxQuant}{(Required). Loaded raw MaxQuant data}

\item{intensityPattern}{(Required). A string of CON, TIME, and REP separated by an underscore, "_". The condition term cannot start with a numeric value and the only acceptable symbols allowed in these three terms are Full-Stop/Period (.) or minus-dash (-). E.g. "CON_TIME_REP".}

\item{filterCon}{(Optional). A list of conditions that you want to process. Only conditions from this list will be included in the downstream analyses.. E.g., c("Con1", "Con2)}

\item{filterTime}{(Optional). A list of time points that you want to process. Only time points within this list will be included in the downstream analyses. E.g., c("0min", "5min", "15min")}

\item{filterRep}{(Optional). A list of replications that you want to process. Only replications from this list will be included in the downstream analyses. E.g., c("R1", "R3", "R4")}
}
\value{
list of data.frames
}
\description{
The function identifies the intensity columns. To correctly identify the columns and for downstream analysis the intensity column names should consist of a string of characters that includes the Condition, Time Point, and Replication or Independent experiment. These three variables need to be separated by an underscore, "_". The function takes as input CON, TIME, REP, separated by an underscore and in the order that these variables are found in the column names. For example, "CON_TIME_REP" will parse the intensity column named "Condition1_0minutes_1".
}
\examples{
\dontrun{
## Loading One Condition Data
data("oneConditionExample")
## Identify the Intensity Columns and Condition, Time Point and Replication
intensityCols <- confirmIntensityColumns(rawMaxQuant = oneConditionExample,
                                     	intensityPattern = "con_time_rep",
                                     	verbose = TRUE)
}
}
