% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractData.R
\name{convertShinyToR}
\alias{convertShinyToR}
\title{Load Shiny Plot Object to R}
\usage{
convertShinyToR(ShinyRDS = ShinyRDS)
}
\arguments{
\item{ShinyRDS}{RDS object downloaded for a plot in NetPhorce Shiny.}
}
\value{
A NetPhorce Object
}
\description{
Load the Shiny Plot Objects to R for more manipulation of colors, fills and layout.
}
