% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractData.R
\name{extractNetworkResults}
\alias{extractNetworkResults}
\title{Extract the network inference results from processed NetPhorce data}
\usage{
extractNetworkResults(netPhorceData = netPhorceData)
}
\arguments{
\item{netPhorceData}{Processed NetPhorce Object}
}
\value{
A single data.frame
}
\description{
Obtain the network inference from NetPhorce data after \code{\link{networkAnalysis}}.
}
\examples{
\dontrun{
## Loading One Condition Data
data("oneConditionExample")
## Identify the Key Columns
identifiedCols <- confirmColumnNames(rawMaxQuant = oneConditionExample,
                                    positionCol = "Position",
                                    reverseCol = "Reverse",
                                    localizationProbCol = "Localization prob",
                                    potentialContaminationCol = "Potential contaminant",
                                    aminoAcidCol = "Amino acid",
                                    uniqueIDCol = "Protein",
                                    seqWindowIDCol = "Sequence window",
                                    fastaIDCol = "Fasta headers")
## Identify the Intensity Columns with Condition, Time Point and Replication Information
intensityCols <- confirmIntensityColumns(rawMaxQuant = oneConditionExample,
                                         intensityPattern = "con_time_rep",
                                         verbose = TRUE)
## Process the data based on the identified columns
netPhorceData <- processData(rawMaxQuant = oneConditionExample,
                             processedColNames = identifiedCols,
                             processedIntensity = intensityCols,
                             minReplication = 3,
                             minLocalProb = 0.75)

## Confirm the correct Kinase/Phosphotase information for the data
netPhorceData <- validateKinaseTable(netPhorceData = netPhorceData,
                                     defaultKinaseTable = TRUE,
                                     abbrev = "Ath")
## Evaluate the netPhorce object with regulation thresholds
netPhorceData <- regulationCheck(netPhorceData = netPhorceData,
                                 upReg = 0.25,
                                 downReg = 0.25,
                                 absMinThreshold = 0.1,
                                 qValueCutOff = 0.05,
                                 verbose = TRUE)
## Network Analysis
netPhorceData <- networkAnalysis(netPhorceData = netPhorceData,
                               requestPlotData = TRUE)
## Extract Network Results
networkResults <- extractNetworkResults(netPhorceData = netPhorceData)
}
}
