% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractData.R
\name{extractSummaryTable}
\alias{extractSummaryTable}
\title{Extract a summary table from processed NetPhorce data}
\usage{
extractSummaryTable(netPhorceData = netPhorceData)
}
\arguments{
\item{netPhorceData}{Processed NetPhorce Object}
}
\value{
A single data.frame
}
\description{
Obtain the summary table from NetPhorce data after \code{\link{processData}}.
}
\examples{
\dontrun{
## Loading One Condition Data
data("oneConditionExample")
## Identify the Key Columns
identifiedCols <- confirmColumnNames(rawMaxQuant = oneConditionExample,
                                    positionCol = "Position",
                                    reverseCol = "Reverse",
                                    localizationProbCol = "Localization prob",
                                    potentialContaminationCol = "Potential contaminant",
                                    aminoAcidCol = "Amino acid",
                                    uniqueIDCol = "Protein",
                                    seqWindowIDCol = "Sequence window",
                                    fastaIDCol = "Fasta headers")
## Identify the Intensity Columns with Condition, Time Point and Replication Information
intensityCols <- confirmIntensityColumns(rawMaxQuant = oneConditionExample,
                                         intensityPattern = "con_time_rep",
                                         verbose = TRUE)
## Process the data based on the identified columns
netPhorceData <- processData(rawMaxQuant = oneConditionExample,
                             processedColNames = identifiedCols,
                             processedIntensity = intensityCols,
                             minReplication = 3,
                             minLocalProb = 0.75)
## Extract Summary Table
summaryTable <- extractSummaryTable(netPhorceData = netPhorceData)
}
}
