% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{findPeptideIDs}
\alias{findPeptideIDs}
\title{Identify provided peptide IDs for plotting}
\usage{
findPeptideIDs(netPhorceData = netPhorceData, peptideIDList = NULL)
}
\arguments{
\item{netPhorceData}{Processed NetPhorce Object}

\item{peptideIDList}{A single or multiple peptide IDs. The peptide ID consists of a combination of the protein ID, amino acid, and multiplicity. If left as \code{NULL}, the full table of available IDs will be printed.}
}
\value{
list of data.frames
}
\description{
Return the input ID when avaiable in the dataset or output all available IDs
}
\examples{
\dontrun{
## Loading Two Conditions Example
data("twoConditionsExample")
## Identify the Key Columns
identifiedCols <- confirmColumnNames(rawMaxQuant = twoConditionsExample,
                                    positionCol = "Position",
                                    reverseCol = "Reverse",
                                    localizationProbCol = "Localization prob",
                                    potentialContaminationCol = "Potential contaminant",
                                    aminoAcidCol = "Amino acid",
                                    uniqueIDCol = "Protein",
                                    seqWindowIDCol = "Sequence window",
                                    fastaIDCol = "Fasta headers")
## Identify the Intensity Columns with Condition, Time Point and Replication Information
intensityCols <- confirmIntensityColumns(rawMaxQuant = twoConditionsExample,
                                         intensityPattern = "con_time_rep",
                                         verbose = TRUE)
## Process the data based on the identified columns
netPhorceData <- processData(rawMaxQuant = twoConditionsExample,
                             processedColNames = identifiedCols,
                             processedIntensity = intensityCols,
                             minReplication = 3,
                             minLocalProb = 0.75)
## Extract peptides from the NetPhorce Data
findPeptideIDs(netPhorceData = netPhorceData,
               peptideIDList = c("AT1G01320.2_S1349_1",
                                 "AT1G05560.1_S153_1",
                                 "AT1G01320.2_S149_1"))
}
}
