% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kinasesPhosphatases}
\alias{kinasesPhosphatases}
\title{Pre-loaded Kinase/Phosphatase Table for 25 Species}
\format{
A data frame with 27,534 kinases and phosphatases records across 25 Species, of which,
\describe{
\item{Abb}{Three-letter abbreviation for the species}
\item{Species}{Full species name}
\item{ID}{Kinase/Phosphatase ID}
\item{Family}{Kinase or Phosphatase classification}
}
}
\usage{
data(kinasesPhosphatases)
}
\description{
The table contains previously published kinase or phosphatase proteins for 25 species. Please use \code{\link{ checkPreloadKinaseTable}} to print a summary for the kinase and phosphatase table.
}
\examples{
data(kinasesPhosphatases)
}
\keyword{datasets}
