% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{netPhorce-class}
\alias{netPhorce-class}
\title{The main class for NetPhorce data}
\description{
Contains the outputs from the phosphoproteome data analysis, statistical analysis, and network inference, and the necessary data for plotting and data extraction.
\describe{
\item{Design}{Contains data design information and filtering parameters}
\item{data.filtered}{Contains the data points that passed filtering criteria in a long format.}
\item{data.filtered.aov.summary}{Contains the  anova results in a long format.}
\item{regulationData}{Contains data associated with regulation check. }
\item{netPhorceResults}{Contains the network inference data. }
\item{networkPlotData}{Contains the data needed for network plotting using \code{\link[visNetwork]{-package}} package}
\item{Misc}{Contains accessory data including default plotting colors and FASTA Keys, if present.}
}
}
