% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oneConditionExample}
\alias{oneConditionExample}
\title{A MaxQuant result file of an experiment with one experimental condition and 5 time points.}
\format{
A data frame with 340 samples and 299 columns, of which,
}
\usage{
data(oneConditionExample)
}
\description{
The demo file is a MaxQuant result file for an experiment with one experimental condition, 5 time points, and 4 replicates. This file contains 340 detected peptides and 299 columns. Only a selection of those columns will be used for the actual data analysis.
}
\examples{
data(oneConditionExample)
}
\keyword{datasets}
