% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiedDEPFunction.R
\name{plot_normalization}
\alias{plot_normalization}
\title{Mofidy DEP function to fit the format of Processed MaxQuant Data}
\usage{
plot_normalization(se, se_norm, filterCondition = NULL)
}
\arguments{
\item{se}{(Required). SummarizedExperiment object from the NetPhorce Object, log2-transformed raw count.}

\item{se_norm}{(Required). SummarizedExperiment object from the NetPhorce Object. normalized log2-transformed raw count.}

\item{filterCondition}{(Required). Experimental condition.}
}
\value{
plot object
}
\description{
The original functions from DEP pacakge can be found at:
https://bioconductor.org/packages/release/bioc/html/DEP.html
The only modification made is replace "condition" with "experiement" to match
the NetPhorce definition of each of these two terms. normalize_vsn and theme_DEP1
also listed here as necessary functions for plot_normalization.
}
\examples{
NULL
}
\keyword{internal}
