% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotNetPhorce}
\alias{plotNetPhorce}
\title{A VisNetwork plot of the inferred network}
\usage{
plotNetPhorce(
  netPhorceData = netPhorceData,
  condition = NULL,
  FASTADescription = FALSE
)
}
\arguments{
\item{netPhorceData}{Processed NetPhorce Object}

\item{condition}{(Required). Select a specific condition if multiple conditions are present.}

\item{FASTADescription}{(Optional). If TRUE, the hover information of the nodes will return the FASTA header column information for better identification of the phosphopeptides.}
}
\value{
A visNetwork object
}
\description{
Return the phosphorylation signaling network with the inferred causal relations between the kinases/phosphatases and downstream peptides.
}
\examples{
\dontrun{
## Loading Two Conditions Example
data("twoConditionsExample")
## Identify the Key Columns
identifiedCols <- confirmColumnNames(rawMaxQuant = twoConditionsExample,
                                    positionCol = "Position",
                                    reverseCol = "Reverse",
                                    localizationProbCol = "Localization prob",
                                    potentialContaminationCol = "Potential contaminant",
                                    aminoAcidCol = "Amino acid",
                                    uniqueIDCol = "Protein",
                                    seqWindowIDCol = "Sequence window",
                                    fastaIDCol = "Fasta headers")
## Identify the Intensity Columns with Condition, Time Point and Replication Information
intensityCols <- confirmIntensityColumns(rawMaxQuant = twoConditionsExample,
                                         intensityPattern = "con_time_rep",
                                         verbose = TRUE)
## Process the data based on the identified columns
netPhorceData <- processData(rawMaxQuant = twoConditionsExample,
                             processedColNames = identifiedCols,
                             processedIntensity = intensityCols,
                             minReplication = 3,
                             minLocalProb = 0.75)
## Confirm the correct Kinase/Phosphotase information for the data
netPhorceData <- validateKinaseTable(netPhorceData = netPhorceData,
                                     defaultKinaseTable = TRUE,
                                     abbrev = "Ath")
## Evaluate the netPhorce object with regulation thresholds
netPhorceData <- regulationCheck(netPhorceData = netPhorceData,
                                 upReg = 0.25,
                                 downReg = 0.25,
                                 absMinThreshold = 0.1,
                                 qValueCutOff = 0.05,
                                 verbose = TRUE)
## Network Analysis
netPhorceData <- networkAnalysis(netPhorceData = netPhorceData,
                               requestPlotData = TRUE)
## VisNetwork Plot with FASTA header information included
plotNetPhorce(netPhorceData = netPhorceData,
              condition = "tot3",
              FASTADescription = TRUE)
}
}
