% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotPCA}
\alias{plotPCA}
\title{PCA Plot}
\usage{
plotPCA(
  netPhorceData = netPhorceData,
  condition = NULL,
  normalized = FALSE,
  plotly = FALSE
)
}
\arguments{
\item{netPhorceData}{(Required). Processed NetPhorce Object}

\item{condition}{(Required). Select a specific condition from your experiment.}

\item{normalized}{(Required). Use raw or normalized intensity data}

\item{plotly}{(Required). If TRUE, output an interactive \code{\link{plotly}} version, else output a static \code{\link{ggplot2}} version.}
}
\value{
A ggplot/plotly object
}
\description{
The principle component analysis (PCA) output before and after variance stabilizing normalization (vsn).
}
\examples{
\dontrun{
## Loading Two Conditions Example
data("twoConditionsExample")
## Identify the Key Columns
identifiedCols <- confirmColumnNames(
  rawMaxQuant = twoConditionsExample,
  positionCol = "Position",
  reverseCol = "Reverse",
  localizationProbCol = "Localization prob",
  potentialContaminationCol = "Potential contaminant",
  aminoAcidCol = "Amino acid",
  uniqueIDCol = "Protein",
  seqWindowIDCol = "Sequence window",
  fastaIDCol = "Fasta headers")
## Identify the Intensity Columns with Condition, Time Point and Replication Information
intensityCols <- confirmIntensityColumns(rawMaxQuant = twoConditionsExample,
                                         intensityPattern = "con_time_rep",
                                         verbose = TRUE)
## Process the data based on the identified columns
netPhorceData <- processData(rawMaxQuant = twoConditionsExample,
                             processedColNames = identifiedCols,
                             processedIntensity = intensityCols,
                             minReplication = 3,
                             minLocalProb = 0.75)
## Generate Distribution Plot – GGPLOT version (Static)
plotPCA(netPhorceData = netPhorceData,
        condition = "tot3",
        plotly = FALSE)
## Generate Distribution Plot – PLOTLY version (Interactive)
plotPCA(netPhorceData = netPhorceData,
        condition = "Col0",
        plotly = TRUE)
}
}
