% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{show,netPhorce-method}
\alias{show,netPhorce-method}
\title{method extensions to show for phyloseq objects.}
\usage{
\S4method{show}{netPhorce}(object)
}
\arguments{
\item{object}{Any R object}
}
\description{
See the general documentation of \code{\link[methods]{show}} method for
expected behavior.
}
\examples{
# data(netPhorceData)
# show(netPhorceData)
# netPhorceData
}
\seealso{
\code{\link[methods]{show}}
}
\keyword{internal}
