% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{twoConditionsExample}
\alias{twoConditionsExample}
\title{A MaxQuant result file of an experiment with two experimental conditions and 5 time points.}
\format{
A data frame with 790 samples and 299 columns, of which,
}
\usage{
data(twoConditionsExample)
}
\description{
The demo file is a MaxQuant result file for an experiment with two experimental conditions (\verb{loss-of-function line} and \verb{control line}), 5 time points, and 4 replicates. The file contains 790 peptides and 673 columns. Only a selection of those columns will be used for the actual data analysis.
}
\examples{
data(twoConditionsExample)
}
\keyword{datasets}
